tox integration of refdata generators
