--- test/quantum_volume/qv_ideal_probabilities.json
@@ -1,1 +1,0 @@
-[[0.06090981922741352, 0.34784828887489516, 0.003440728125794299, 0.17695071101919052, 0.020399173510863133, 0.3133909057297853, 0.025208906876148903, 0.051851466635908984], [0.052795588044998266, 0.2036758692494191, 0.022228933273601266, 0.267947067954847, 0.04662254705199523, 0.23689106637344887, 0.014644198017660856, 0.15519473003402912], [0.33890568849498903, 0.08470251706991236, 0.09032425795803736, 0.1176818050135521, 0.20023444940934337, 0.040376649609833996, 0.07845708684381617, 0.04931754560051594], [0.07843754575087986, 0.21907218041759557, 0.11588111917837404, 0.14344927568779245, 0.07247077881185233, 0.13642212295586695, 0.19928429396422193, 0.03498268323341709], [0.03146764504989369, 0.010672855981310281, 0.0698472783532379, 0.2805983021183489, 0.1697510376611356, 0.34490526483543, 0.05675488127108915, 0.03600273472955497], [0.014712440805189736, 0.20198281763070447, 0.03951483459147206, 0.10661868457522322, 0.028973459527277957, 0.0030555802881049895, 0.1683968533591363, 0.43674532922289167], [0.4652838497088344, 0.1370319640546792, 0.0, 0.0, 0.11184911640086283, 0.2858350698356241, 0.0, 0.0], [0.41691761271850697, 0.09802206331301935, 0.0, 0.0, 0.12170171915272401, 0.36335860481574933, 0.0, 0.0], [0.06040824842131547, 0.22360037606154226, 0.16703416728799328, 0.3155515725112804, 0.00574389769671973, 0.06422576739914752, 0.05653984347944168, 0.10689612714255917], [0.034145937220083696, 0.3275471371847169, 0.0, 0.0, 0.5265934908959305, 0.11171343469926844, 0.0, 0.0], [0.31362955944335913, 0.0012492306942707208, 0.0, 0.0, 0.3045171469773979, 0.380604062884972, 0.0, 0.0], [0.14084298473415383, 0.04241815646129184, 0.2787065997143041, 0.2763086311775977, 0.04767796566311971, 0.04777093880194504, 0.030317360310517225, 0.1359573631370707], [0.10490558188826622, 0.03734373973356664, 0.0846661393845946, 0.012178280925959687, 0.2450469596449601, 0.3830889806436624, 0.0410475620559563, 0.09172275572303307], [0.0721384781481662, 0.2132011346169562, 0.0382461870322736, 0.01006147040056322, 0.0012564152657445, 0.5640473347441018, 0.03581099593386722, 0.06523798385832678], [0.35047461795366364, 0.07295546107526876, 0.11429248870690438, 0.046387738341356205, 0.016403322359457823, 0.11703328502400306, 0.18840822678605898, 0.09404485975328614], [0.12537608375594167, 0.004147254265659097, 0.01888208426232345, 0.127828626942106, 0.20712391654423107, 0.004956839798677847, 0.20103868882942894, 0.31064650560163193], [0.24266832167544544, 0.10171085169237129, 0.6518557453504638, 0.0037650812817186354, 0.0, 0.0, 0.0, 0.0], [0.004496439754867754, 0.03999764295525122, 0.045209067627212385, 0.42243909053043943, 0.23660227820571852, 0.1145783400317101, 0.016512735149122833, 0.12016440574567633], [0.403998318316775, 0.1913339052666593, 0.39512059243880765, 0.009547183977757986, 0.0, 0.0, 0.0, 0.0], [0.046975692653985185, 0.07680806404698703, 0.013699720588740898, 0.1007536179821163, 0.006010045196316757, 0.49827221242540065, 0.0039712903211504315, 0.25350935678530284]]
