--- test/quantum_volume/qv_result_moderate_noise_300_trials.json
@@ -1,1 +1,0 @@
-[{"name": "mean_HOP", "value": {"__type__": "__object__", "__value__": {"__name__": "FitVal", "__module__": "qiskit_experiments.database_service.db_fitval", "__kwargs__": {"value": 0.73146484375, "stderr": 0.025588019729799863, "unit": null}}}, "extra": {"HOPs": [0.7587890625, 0.77734375, 0.751953125, 0.7353515625, 0.8017578125, 0.630859375, 0.75, 0.7783203125, 0.7763671875, 0.7744140625, 0.7099609375, 0.8291015625, 0.7109375, 0.7119140625, 0.6884765625, 0.7890625, 0.75, 0.69921875, 0.7216796875, 0.6689453125, 0.6826171875, 0.693359375, 0.650390625, 0.7255859375, 0.7451171875, 0.7529296875, 0.8876953125, 0.7578125, 0.7041015625, 0.720703125, 0.6884765625, 0.734375, 0.814453125, 0.7841796875, 0.6904296875, 0.6640625, 0.78125, 0.7041015625, 0.689453125, 0.69140625, 0.8134765625, 0.7119140625, 0.7373046875, 0.7099609375, 0.7890625, 0.6572265625, 0.7734375, 0.7353515625, 0.744140625, 0.8447265625, 0.6748046875, 0.734375, 0.66796875, 0.689453125, 0.7080078125, 0.767578125, 0.703125, 0.7958984375, 0.6845703125, 0.787109375, 0.76171875, 0.7333984375, 0.7685546875, 0.8046875, 0.7265625, 0.732421875, 0.7802734375, 0.7529296875, 0.765625, 0.759765625, 0.732421875, 0.7431640625, 0.7900390625, 0.6875, 0.6474609375, 0.7236328125, 0.708984375, 0.6064453125, 0.71484375, 0.740234375, 0.7392578125, 0.7255859375, 0.779296875, 0.681640625, 0.6826171875, 0.6669921875, 0.708984375, 0.7109375, 0.7158203125, 0.6943359375, 0.7763671875, 0.73828125, 0.673828125, 0.791015625, 0.748046875, 0.7685546875, 0.7392578125, 0.6923828125, 0.7685546875, 0.7646484375, 0.763671875, 0.7568359375, 0.7314453125, 0.7783203125, 0.7705078125, 0.765625, 0.73046875, 0.81640625, 0.744140625, 0.7158203125, 0.751953125, 0.720703125, 0.6923828125, 0.6875, 0.7021484375, 0.6728515625, 0.6748046875, 0.6455078125, 0.73046875, 0.724609375, 0.740234375, 0.6455078125, 0.8193359375, 0.69921875, 0.7578125, 0.798828125, 0.7919921875, 0.701171875, 0.72265625, 0.7021484375, 0.7578125, 0.794921875, 0.724609375, 0.7109375, 0.685546875, 0.8388671875, 0.7119140625, 0.7548828125, 0.6904296875, 0.6943359375, 0.6220703125, 0.7333984375, 0.7177734375, 0.75390625, 0.6611328125, 0.7294921875, 0.783203125, 0.80078125, 0.701171875, 0.765625, 0.6884765625, 0.6845703125, 0.751953125, 0.693359375, 0.7294921875, 0.8349609375, 0.697265625, 0.736328125, 0.732421875, 0.7041015625, 0.673828125, 0.759765625, 0.7626953125, 0.75, 0.72265625, 0.7666015625, 0.705078125, 0.6982421875, 0.751953125, 0.7421875, 0.6689453125, 0.78125, 0.7880859375, 0.76171875, 0.71875, 0.6904296875, 0.744140625, 0.7578125, 0.7412109375, 0.6806640625, 0.80078125, 0.6982421875, 0.724609375, 0.7275390625, 0.6552734375, 0.66015625, 0.6884765625, 0.771484375, 0.736328125, 0.6513671875, 0.794921875, 0.6748046875, 0.7041015625, 0.783203125, 0.712890625, 0.8359375, 0.7412109375, 0.7236328125, 0.703125, 0.6689453125, 0.73046875, 0.7998046875, 0.751953125, 0.6982421875, 0.7412109375, 0.740234375, 0.736328125, 0.6484375, 0.72265625, 0.751953125, 0.7919921875, 0.7333984375, 0.7021484375, 0.740234375, 0.76953125, 0.7421875, 0.6953125, 0.70703125, 0.7236328125, 0.736328125, 0.7060546875, 0.7265625, 0.79296875, 0.76953125, 0.669921875, 0.806640625, 0.8076171875, 0.69921875, 0.7333984375, 0.712890625, 0.8798828125, 0.712890625, 0.73046875, 0.720703125, 0.701171875, 0.6865234375, 0.7265625, 0.7529296875, 0.68359375, 0.6533203125, 0.7138671875, 0.666015625, 0.7724609375, 0.7138671875, 0.8232421875, 0.701171875, 0.724609375, 0.7255859375, 0.7578125, 0.7333984375, 0.693359375, 0.767578125, 0.720703125, 0.734375, 0.7021484375, 0.7265625, 0.6884765625, 0.740234375, 0.6845703125, 0.6884765625, 0.7412109375, 0.7001953125, 0.6953125, 0.6669921875, 0.798828125, 0.7197265625, 0.7431640625, 0.7685546875, 0.73046875, 0.7119140625, 0.7119140625, 0.7275390625, 0.80859375, 0.68359375, 0.7314453125, 0.6728515625, 0.7041015625, 0.7333984375, 0.6884765625, 0.740234375, 0.7587890625, 0.701171875, 0.75390625, 0.662109375, 0.7685546875, 0.7841796875, 0.7646484375, 0.748046875, 0.7216796875, 0.7041015625, 0.7568359375, 0.736328125, 0.80859375, 0.7509765625, 0.8125, 0.75390625, 0.69921875, 0.712890625, 0.6748046875, 0.755859375], "two_sigma": 0.051176039459599726, "depth": 4, "trials": 300}}, {"name": "quantum_volume", "value": 16, "extra": {"success": true, "confidence": 0.9943351826324864, "depth": 4, "trials": 300}}]
