+++ test/quantum_volume/refdata/data_high_confidence.py
@@ -1,0 +1,90 @@
+# This code is part of Qiskit.
+#
+# (C) Copyright IBM 2021.
+#
+# This code is licensed under the Apache License, Version 2.0. You may
+# obtain a copy of this license in the LICENSE.txt file in the root directory
+# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# Any modifications or derivative works of this code must retain this
+# copyright notice, and modified files need to carry a notice indicating
+# that they have been altered from the originals.
+"""
+High confidence data generation.
+"""
+import argparse
+import json
+import os
+
+from qiskit import Aer
+
+from qiskit_experiments.database_service.json import ExperimentEncoder
+from qiskit_experiments.library import QuantumVolume
+
+try:
+    from .utils import create_noise_model
+except ImportError:
+    # we cannot use relative import from the top level module.
+    from utils import create_noise_model
+
+
+def _create_qv_data_high_confidence(dir_path: str, seed_val: int):
+    """
+    create quantum volume experiment_data using seed, and save it as a json
+    the circuit is generated with moderate noise, so the mean hop is above 2/3,
+    and also with enough trials, so the confidence is above threshold
+
+    Args:
+        dir_path(str): The directory which the data will be saved to.
+        seed_val (int): Seed value.
+    """
+    num_of_qubits = 4
+    backend = Aer.get_backend("aer_simulator")
+    basis_gates = ["id", "rz", "sx", "x", "cx", "reset"]
+    noise = create_noise_model()
+
+    qv_exp = QuantumVolume(num_of_qubits, seed=seed_val)
+    qv_exp.set_experiment_options(trials=300)
+    qv_exp.set_transpile_options(basis_gates=basis_gates)
+    qv_data = qv_exp.run(backend, noise_model=noise, basis_gates=basis_gates)
+    qv_data.block_for_results()
+
+    result_file_path = os.path.join(dir_path, "qv_data_moderate_noise_300_trials.json")
+    with open(result_file_path, "w") as json_file:
+        json.dump(qv_data.data(), json_file, cls=ExperimentEncoder)
+
+    result_file_path = os.path.join(dir_path, "qv_result_moderate_noise_300_trials.json")
+    with open(result_file_path, "w") as json_file:
+        result_dicts = []
+        for result in qv_data.analysis_results():
+            result_dicts.append(
+                {
+                    "name": result.name,
+                    "value": result.value,
+                    "extra": result.extra,
+                }
+            )
+        json.dump(result_dicts, json_file, cls=ExperimentEncoder)
+
+
+if __name__ == "__main__":
+    parser = argparse.ArgumentParser(description='Quantum volume ref data generation.')
+    parser.add_argument(
+        '--folder',
+        required=False,
+        default="test/quantum_volume/refdata",
+        type=str,
+    )
+    parser.add_argument(
+        '--seed',
+        required=False,
+        default=42,
+        type=int,
+    )
+    args = parser.parse_args()
+
+    print(f"Generating data for {os.path.basename(__file__)}...")
+
+    _create_qv_data_high_confidence(args.folder, args.seed)
+
+    print(f"Completed")
