--- test/randomized_benchmarking/rb_interleaved_1qubit_output_analysis.json
@@ -1,1 +1,0 @@
-[{"name": "@Parameters_InterleavedRBAnalysis", "value": {"__type__": "__object__", "__value__": {"__name__": "FitVal", "__module__": "qiskit_experiments.database_service.db_fitval", "__kwargs__": {"value": {"__type__": "array", "__value__": [0.5116650227730736, 0.9994768276408555, 0.9979725591597532, 0.4884021095472425]}, "stderr": {"__type__": "array", "__value__": [0.006437954091857918, 1.4847358402650982e-05, 6.882915444481367e-05, 0.006675424355241532]}, "unit": null}}}, "extra": {"popt_keys": ["a", "alpha", "alpha_c", "b"], "dof": 16, "covariance_mat": {"__type__": "array", "__value__": [[4.144725288887011e-05, 4.533933132185307e-08, 3.746653829699993e-07, -4.023979858687134e-05], [4.5339331321853074e-08, 2.2044405153677072e-10, 6.432549372321611e-10, -6.928087572143161e-08], [3.746653829699994e-07, 6.432549372321611e-10, 4.737452501588013e-09, -4.2638270013676844e-07], [-4.023979858687135e-05, -6.92808757214316e-08, -4.2638270013676844e-07, 4.456129032255182e-05]]}}}, {"name": "alpha", "value": {"__type__": "__object__", "__value__": {"__name__": "FitVal", "__module__": "qiskit_experiments.database_service.db_fitval", "__kwargs__": {"value": 0.9994768276408555, "stderr": 1.4847358402650982e-05, "unit": null}}}, "extra": {}}, {"name": "alpha_c", "value": {"__type__": "__object__", "__value__": {"__name__": "FitVal", "__module__": "qiskit_experiments.database_service.db_fitval", "__kwargs__": {"value": 0.9979725591597532, "stderr": 6.882915444481367e-05, "unit": null}}}, "extra": {}}, {"name": "EPC", "value": {"__type__": "__object__", "__value__": {"__name__": "FitVal", "__module__": "qiskit_experiments.database_service.db_fitval", "__kwargs__": {"value": 0.0010137204201233763, "stderr": 3.4414577222406835e-05, "unit": null}}}, "extra": {"EPC_systematic_err": 0.0010137204201233763, "EPC_systematic_bounds": [0.0, 0.0020274408402467525]}}]
