--- test/randomized_benchmarking/rb_interleaved_2qubits_output_analysis.json
@@ -1,1 +1,0 @@
-[{"name": "@Parameters_InterleavedRBAnalysis", "value": {"__type__": "__object__", "__value__": {"__name__": "FitVal", "__module__": "qiskit_experiments.database_service.db_fitval", "__kwargs__": {"value": {"__type__": "array", "__value__": [0.7342127078996477, 0.9884063217136179, 0.9927651616883224, 0.2543568492711422]}, "stderr": {"__type__": "array", "__value__": [0.004948839577232964, 0.00020370367690368603, 0.00044396411042263065, 0.004356375518781768]}, "unit": null}}}, "extra": {"popt_keys": ["a", "alpha", "alpha_c", "b"], "dof": 16, "covariance_mat": {"__type__": "array", "__value__": [[2.4491013161187343e-05, 5.236454920804876e-07, 1.2793043417064756e-06, -1.6442242444492406e-05], [5.236454920804876e-07, 4.149518798408131e-08, 4.9409520508775484e-08, -7.419439279427566e-07], [1.2793043417064754e-06, 4.940952050877548e-08, 1.971041313433578e-07, -1.5342339511770156e-06], [-1.6442242444492406e-05, -7.419439279427566e-07, -1.5342339511770158e-06, 1.8978007660641124e-05]]}}}, {"name": "alpha", "value": {"__type__": "__object__", "__value__": {"__name__": "FitVal", "__module__": "qiskit_experiments.database_service.db_fitval", "__kwargs__": {"value": 0.9884063217136179, "stderr": 0.00020370367690368603, "unit": null}}}, "extra": {}}, {"name": "alpha_c", "value": {"__type__": "__object__", "__value__": {"__name__": "FitVal", "__module__": "qiskit_experiments.database_service.db_fitval", "__kwargs__": {"value": 0.9927651616883224, "stderr": 0.00044396411042263065, "unit": null}}}, "extra": {}}, {"name": "EPC", "value": {"__type__": "__object__", "__value__": {"__name__": "FitVal", "__module__": "qiskit_experiments.database_service.db_fitval", "__kwargs__": {"value": 0.005426128733758195, "stderr": 0.000332973082816973, "unit": null}}}, "extra": {"EPC_systematic_err": 0.01196438869581501, "EPC_systematic_bounds": [0, 0.017390517429573205]}}]
