--- test/randomized_benchmarking/rb_standard_1qubit_output_analysis.json
@@ -1,1 +1,0 @@
-[{"name": "@Parameters_RBAnalysis", "value": {"__type__": "__object__", "__value__": {"__name__": "FitVal", "__module__": "qiskit_experiments.database_service.db_fitval", "__kwargs__": {"value": {"__type__": "array", "__value__": [0.9967817420311399, 0.9997606942463159, 2.813384952009873e-13]}, "stderr": {"__type__": "array", "__value__": [1.5428416345109068, 0.00041321430695411904, 1.5480528890603416]}, "unit": null}}}, "extra": {"popt_keys": ["a", "alpha", "b"], "dof": 7, "covariance_mat": {"__type__": "array", "__value__": [[2.3803603091802867, 0.0006371615470064913, -2.388380804138054], [0.0006371615470064913, 1.7074606347157292e-07, -0.0006393835737758766], [-2.388380804138054, -0.0006393835737758766, 2.3964677473280704]]}}}, {"name": "alpha", "value": {"__type__": "__object__", "__value__": {"__name__": "FitVal", "__module__": "qiskit_experiments.database_service.db_fitval", "__kwargs__": {"value": 0.9997606942463159, "stderr": 0.00041321430695411904, "unit": null}}}, "extra": {}}, {"name": "EPC", "value": {"__type__": "__object__", "__value__": {"__name__": "FitVal", "__module__": "qiskit_experiments.database_service.db_fitval", "__kwargs__": {"value": 0.00011965287684206904, "stderr": 0.00020660715347705952, "unit": null}}}, "extra": {}}, {"name": "EPG_x", "value": {"__type__": "__object__", "__value__": {"__name__": "FitVal", "__module__": "qiskit_experiments.database_service.db_fitval", "__kwargs__": {"value": 0.0005083368294731997, "stderr": null, "unit": null}}}, "extra": {}}]
