--- test/randomized_benchmarking/rb_standard_2qubits_output_analysis.json
@@ -1,1 +1,0 @@
-[{"name": "@Parameters_RBAnalysis", "value": {"__type__": "__object__", "__value__": {"__name__": "FitVal", "__module__": "qiskit_experiments.database_service.db_fitval", "__kwargs__": {"value": {"__type__": "array", "__value__": [0.7384628327482253, 0.988240664113885, 0.2440050521435873]}, "stderr": {"__type__": "array", "__value__": [0.014085818516668487, 0.0005254263962064677, 0.015295944033692057]}, "unit": null}}}, "extra": {"popt_keys": ["a", "alpha", "b"], "dof": 7, "covariance_mat": {"__type__": "array", "__value__": [[0.00019841028328452083, 6.157466444295669e-06, -0.00020381596731956612], [6.1574664442956684e-06, 2.760728978305159e-07, -7.599447787142353e-06], [-0.00020381596731956612, -7.5994477871423535e-06, 0.00023396590388183962]]}}}, {"name": "alpha", "value": {"__type__": "__object__", "__value__": {"__name__": "FitVal", "__module__": "qiskit_experiments.database_service.db_fitval", "__kwargs__": {"value": 0.988240664113885, "stderr": 0.0005254263962064677, "unit": null}}}, "extra": {}}, {"name": "EPC", "value": {"__type__": "__object__", "__value__": {"__name__": "FitVal", "__module__": "qiskit_experiments.database_service.db_fitval", "__kwargs__": {"value": 0.008819501914586247, "stderr": 0.00039406979715485074, "unit": null}}}, "extra": {}}, {"name": "EPG_cx", "value": {"__type__": "__object__", "__value__": {"__name__": "FitVal", "__module__": "qiskit_experiments.database_service.db_fitval", "__kwargs__": {"value": 0.005560563562776086, "stderr": null, "unit": null}}}, "extra": {}}]
