+++ tox.ini
@@ -28,5 +28,16 @@ commands = black {posargs} qiskit_experiments test
 commands =
   sphinx-build -b html -W {posargs} docs/ docs/_build/html
 
+[testenv:refdata]
+description = 'Generate reference data for unittest'
+commands =
+  python test/randomized_benchmarking/refdata/data_standard_rb.py
+  python test/randomized_benchmarking/refdata/data_interleaved_rb.py
+  python test/quantum_volume/refdata/data_70_trials.py
+  python test/quantum_volume/refdata/data_high_confidence.py
+  python test/quantum_volume/refdata/data_ideal_probabiities.py
+  python test/quantum_volume/refdata/data_low_confidence.py
+  python test/quantum_volume/refdata/data_low_hop.py
+
 [pycodestyle]
 max-line-length = 100
